requirements.txt
pandas==2.3.3
psycopg==3.2.3
python-dotenv==1.2.1
vault_cli==3.1.0
